Conflicts:
	src/client.ts
	src/resource.ts
	src/types/meta.ts

+++ src/resource.ts
@@ -1,5 +1,5 @@
-import {KubernetesRESTClient} from "./client";
-import {APIObject, MetadataObject, WatchEvent} from "./types/meta";
+import {IKubernetesRESTClient} from "./client";
+import {APIObject, DeleteOptions, MetadataObject, WatchEvent} from "./types/meta";
 import {LabelSelector} from "./label";
 
 export interface WatchHandler<O> {
@@ -9,11 +9,27 @@ export interface WatchHandler<O> {
     onError?(err: any): any;
 }
 
-export class ResourceClient<R extends MetadataObject, K, V> {
+export interface IResourceClient<R extends MetadataObject, K, V> {
+    list(labelSelector?: LabelSelector): Promise<Array<APIObject<K, V> & R>>;
+    get(name: string): Promise<(APIObject<K, V> & R) | undefined>;
+    apply(resource: R): Promise<APIObject<K, V> & R>;
+    put(resource: R): Promise<APIObject<K, V> & R>;
+    post(resource: R): Promise<APIObject<K, V> & R>;
+    delete(resourceOrName: R|string, deleteOptions?: DeleteOptions): Promise<void>;
+    deleteMany(labelSelector: LabelSelector): Promise<void>;
+}
+
+export interface INamespacedResourceClient<R extends MetadataObject, K, V> extends IResourceClient<R, K, V> {
+    namespace(ns: string): IResourceClient<R, K, V>;
+    allNamespaces(): IResourceClient<R, K, V>;
+}
+
+export class ResourceClient<R extends MetadataObject, K, V> implements IResourceClient<R, K, V> {
 
     protected baseURL: string;
+    public supportsCollectionDeletion: boolean = true;
 
-    public constructor(protected client: KubernetesRESTClient,
+    public constructor(protected client: IKubernetesRESTClient,
                        protected apiBaseURL: string,
                        protected resourceBaseURL: string) {
         apiBaseURL = apiBaseURL.replace(/\/$/, "");
@@ -58,7 +74,7 @@ export class ResourceClient<R extends MetadataObject, K, V> {
         return await this.client.post(this.baseURL, resource);
     }
 
-    public async delete(resourceOrName: R|string): Promise<void> {
+    public async delete(resourceOrName: R|string, deleteOptions?: DeleteOptions): Promise<void> {
         let url;
         if (typeof resourceOrName === "string") {
             url = this.baseURL + "/" + resourceOrName;
@@ -66,19 +82,24 @@ export class ResourceClient<R extends MetadataObject, K, V> {
             url = this.urlForResource(resourceOrName);
         }
 
-        return await this.client.delete(url);
+        return await this.client.delete(url, undefined, undefined, deleteOptions);
     }
 
     public async deleteMany(labelSelector: LabelSelector) {
-        return await this.client.delete(this.baseURL, labelSelector);
+        if (this.supportsCollectionDeletion) {
+            return await this.client.delete(this.baseURL, labelSelector);
+        }
+
+        const resources = await this.list(labelSelector);
+        await Promise.all(resources.map(r => this.delete(r)));
     }
 
 }
 
-export class NamespacedResourceClient<R extends MetadataObject, K, V> extends ResourceClient<R, K, V> {
+export class NamespacedResourceClient<R extends MetadataObject, K, V> extends ResourceClient<R, K, V> implements INamespacedResourceClient<R, K, V> {
     private ns?: string;
 
-    public constructor(client: KubernetesRESTClient,
+    public constructor(client: IKubernetesRESTClient,
                        apiBaseURL: string,
                        resourceBaseURL: string,
                        ns?: string) {
@@ -97,15 +118,24 @@ export class NamespacedResourceClient<R extends MetadataObject, K, V> extends Re
     }
 
     protected urlForResource(r: R): string {
-        return this.apiBaseURL + "/namespaces/" + r.metadata.namespace + "/" + this.resourceBaseURL + "/" + r.metadata.name;
+        const namespace = r.metadata.namespace || this.ns;
+        if (namespace) {
+            return this.apiBaseURL + "/namespaces/" + namespace + "/" + this.resourceBaseURL + "/" + r.metadata.name;
+        }
+
+        return this.apiBaseURL + "/" + this.resourceBaseURL + "/" + r.metadata.name;
     }
 
-    public namespace(ns: string): ResourceClient<R, K, V> {
-        return new NamespacedResourceClient<R, K, V>(this.client, this.apiBaseURL, this.resourceBaseURL, ns);
+    public namespace(ns: string): IResourceClient<R, K, V> {
+        const n = new NamespacedResourceClient<R, K, V>(this.client, this.apiBaseURL, this.resourceBaseURL, ns);
+        n.supportsCollectionDeletion = this.supportsCollectionDeletion;
+        return n;
     }
 
-    public allNamespaces(): ResourceClient<R, K, V> {
-        return new NamespacedResourceClient<R, K, V>(this.client, this.apiBaseURL, this.resourceBaseURL);
+    public allNamespaces(): IResourceClient<R, K, V> {
+        const n = new NamespacedResourceClient<R, K, V>(this.client, this.apiBaseURL, this.resourceBaseURL);
+        n.supportsCollectionDeletion = this.supportsCollectionDeletion;
+        return n;
     }
 
     public async post(resource: R) {
